:sparkles: Add a central location for feeling names
The idea here is I want lots of different words to be available to the CLI
as ways to describe the 5 different levels of the feeling scale. What I've
got here isn't exhaustive, and it can and likely will be expanded over time,
and doing it this way I guess also allows for some degree of localisation
should it ever be necessary.