:hammer: Move away from a monolithic file and into individual files
Still not sure, still dialling this in. It think I prefer this approach, I
think it might be slower in some places but have benefits in others. I'm
going to need to think carefully about sorting the data as it's loaded, when
I want to use a whole collection, but I like this more granular approach.

In theory it could even make it more sync-friendly, if using a cloud storage
setup (and if I added some sort of filesystem watcher for the UI or
something).

But, anyway... early days; still thinking about how nest to approach this.